fix: jscon_size() now returning proper type (size_t)

+++ include/libjscon.h
@@ -154,7 +154,7 @@ char* jscon_stringify(jscon_item_st *root, jscon_type_et type);
 void jscon_list_append(jscon_list_st *list, jscon_item_st *item);
 
 /* JSCON UTILITIES */
-long jscon_size(const jscon_item_st* kItem);
+size_t jscon_size(const jscon_item_st* kItem);
 jscon_item_st* jscon_append(jscon_item_st *item, jscon_item_st *new_branch);
 jscon_item_st* jscon_dettach(jscon_item_st *item);
 void jscon_delete(jscon_item_st *item, const char *kKey);
